+++ src/engine/Collision/BoundingBox.ts
@@ -97,6 +97,14 @@ export class BoundingBox {
     return new BoundingBox(minX, minY, maxX, maxY);
   }
 
+  /**
+   * Creates a bounding box from a width and height
+   * @param width
+   * @param height
+   * @param anchor Default Vector.Half
+   * @param pos Default Vector.Zero
+   * @returns
+   */
   public static fromDimension(width: number, height: number, anchor: Vector = Vector.Half, pos: Vector = Vector.Zero) {
     return new BoundingBox(
       -width * anchor.x + pos.x,
@@ -327,7 +335,6 @@ export class BoundingBox {
    * Test wether this bounding box intersects with another returning
    * the intersection vector that can be used to resolve the collision. If there
    * is no intersection null is returned.
-   *
    * @param other  Other [[BoundingBox]] to test intersection with
    * @returns A Vector in the direction of the current BoundingBox, this <- other
    */