+++ CHANGELOG.md
@@ -20,11 +20,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fixed issue with input mapper where `keyboard.wasPressed(...)` did not fire
+- Fixed issue issue where TileMaps would not properly draw Tiles when setup in screen space coordinates
 - Fixed issue where the ex.Line graphics bounds were incorrect causing erroneous offscreen culling
 
 ### Updates
 
--
+- Improved performance in TileMaps when drawing tiles using QuadTree data structure
 
 ### Changed